+++ GameFrame/Controllers/GamePadButton.cs
@@ -1,0 +1,20 @@
+﻿using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Input;
+
+namespace GameFrame.Controllers
+{
+    public class GamePadButton : AbstractGamePadButton
+    {
+        public GamePadButton(Buttons button, PlayerIndex player = PlayerIndex.One)
+        {
+            Button = button;
+            Player = player;
+        }
+
+        public override void Update(GamePadState state)
+        {
+            PreviouslyActive = Active;
+            Active = state.IsButtonDown(Button);
+        }
+    }
+}
